# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Job Portal REST API",
    "version": "1.0.0",
    "description": "Production-grade RESTful API for Job Portal featuring role-based access control, AI-powered matching, and candidate pipelines."
  },
  "servers": [
    {
      "url": "/api",
      "description": "Relative API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/jobs": {
      "get": {
        "summary": "List active jobs with filters and pagination",
        "tags": ["Jobs"],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer" } },
          { "name": "per_page", "in": "query", "schema": { "type": "integer" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "category", "in": "query", "schema": { "type": "string" } },
          { "name": "location", "in": "query", "schema": { "type": "string" } },
          { "name": "job_type", "in": "query", "schema": { "type": "string" } },
          { "name": "sort", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "List of jobs" }
        }
      },
      "post": {
        "summary": "Post a new job listing (Employers/Admins)",
        "tags": ["Jobs"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "company_name", "location", "description"],
                "properties": {
                  "title": { "type": "string" },
                  "company_name": { "type": "string" },
                  "location": { "type": "string" },
                  "job_type": { "type": "string" },
                  "category": { "type": "string" },
                  "description": { "type": "string" },
                  "salary_min": { "type": "number" },
                  "salary_max": { "type": "number" },
                  "skills": { "type": "array", "items": { "type": "string" } }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Job created" },
          "403": { "description": "Forbidden" }
        }
      }
    },
...
    "/admin/security/overview": {
      "get": {
        "summary": "Admin security telemetry and report statistics",
        "tags": ["Admin Security"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Security overview metrics" }
        }
      }
    }
  }
}